pyproject.toml
[project]
name = "prepnexa-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "boto3>=1.42.14",
    "celery>=5.6.0",
    "fastapi[standard]>=0.125.0",
    "openai>=2.14.0",
    "opencv-python>=4.12.0.88",
    "passlib>=1.7.4",
    "pdf2image>=1.17.0",
    "pillow>=12.0.0",
    "pydantic-settings>=2.12.0",
    "pytesseract>=0.3.13",
    "python-jose>=3.5.0",
    "redis>=7.1.0",
    "reportlab>=4.4.7",
    "sqlalchemy>=2.0.45",
]
